src/mod-3-tex-to-sql/src/mindsdb/requirements.txt
mindsdb_sdk>=1.8.0
streamlit>=1.28.0
python-dotenv>=1.0.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
requests>=2.31.0